Move KernelTree into its own file
